Version 0.2
Initialized PWM for two servos.  GPIO12 and GPIO18 on a RPi 3 board will
be used to control the gates.  Both gate servos will start at the full
left position which will be gates down.  The plan is to lift one gate
before entering the polling loop for the button inputs.